requirements.txt
sqlalchemy>=2.0,<3.0
pypdf>=4.2,<5.0
openpyxl>=3.1,<4.0
python-dotenv>=1.0,<2.0
langgraph>=0.2,<1.0
streamlit>=1.37,<2.0
requests>=2.32,<3.0
pymupdf>=1.24,<2.0
